requirements.txt
YOUTUBE_TRANSCRIPT_API
streamlit
google-generativeai
python-dotenv
pathlib